LayoutTests/media/track/video/video-track-mkv-theora-id-expected.txt


Check video tracks' language attributes.

EVENT(canplaythrough)

** Check in-band kind attributes
EXPECTED (video.videoTracks.length == '2') OK
EXPECTED (video.videoTracks[0].id == '12184918091563849224') OK
EXPECTED (video.videoTracks[1].id == '4514952819552700788') OK

END OF TEST
